Separete platform specific code to separate files
