+++ src/main.c
@@ -4,27 +4,11 @@
 // Author: Shihua (Simon) Xiao, sixiao@microsoft.com
 // ----------------------------------------------------------------------------------
 
-#include "main.h"
-
-#ifdef _WIN32
-static int gettimeofday(struct timeval * tp, struct timezone * tzp)
-{
-	static const uint64_t EPOCH = ((uint64_t) 116444736000000000ULL);
-
-	SYSTEMTIME  system_time;
-	FILETIME    file_time;
-	uint64_t    time;
-
-	GetSystemTime( &system_time );
-	SystemTimeToFileTime( &system_time, &file_time );
-	time =  ((uint64_t)file_time.dwLowDateTime );
-	time += ((uint64_t)file_time.dwHighDateTime) << 32;
-
-	tp->tv_sec  = (long) ((time - EPOCH) / 10000000L);
-	tp->tv_usec = (long) (system_time.wMilliseconds * 1000);
-        return 0;
-}
-#endif
+#include "common.h"
+#include "util.h"
+#include "logger.h"
+#include "tcpstream.h"
+#include "controller.h"
 
 static int n_write_read(int sockfd, char *buffer, int msg_actual_size)
 {
