--- src/main.h
@@ -1,17 +1,0 @@
-// ----------------------------------------------------------------------------------
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-// Author: Shihua (Simon) Xiao, sixiao@microsoft.com
-// ----------------------------------------------------------------------------------
-
-#ifndef MAIN_H
-#define MAIN_H
-
-#define _GNU_SOURCE
-
-#include "common.h"
-#include "util.h"
-#include "tcpstream.h"
-#include "controller.h"
-
-#endif // MAIN_H
